fix: 리뷰토탈 수정
