+++ client/src/components/MyActivity.jsx
@@ -11,11 +11,7 @@ function MyActivity() {
 
   const [currentTab, setCurrentTab] = useState(0);
   const [reviewCount, setRieviewCount] = useState(0);
-  const [target, setTarget] = useState(null);
-  const [isLoaded, setIsLoaded] = useState(false);
-  let page = 1;
-  let total = 1;
-  let lock = false;
+
   const [reviews, setReviews] = useState([]);
 
   const menuArr = [
@@ -26,14 +22,7 @@ function MyActivity() {
     { name: '|' },
     {
       name: `내가 쓴 리뷰 (${reviewCount})`,
-      content: (
-        <MyReviews
-          isLoaded={isLoaded}
-          reviews={reviews}
-          setReviews={setReviews}
-          setTarget={setTarget}
-        />
-      ),
+      content: <MyReviews reviews={reviews} setReviews={setReviews} />,
     },
   ];
 
@@ -44,54 +33,9 @@ function MyActivity() {
       loading: false,
     }).then((res) => {
       setReviews(res.data.items);
-      setRieviewCount(res.data.pages.total);
-      console.log('내가 쓴 리뷰 토탈', res.data.pages.total);
+      setRieviewCount(res.data.pages.itemsCount);
     });
   }, []);
-
-  //! 더보기
-  const getMoreItem = async () => {
-    if (page > total) {
-      setIsLoaded(false);
-      return true;
-    }
-
-    const res = await axios({
-      url: `/reviews?size=5&page=${page + 1}`,
-      loading: false,
-    });
-    setReviews((review) => review.concat(res.data.items));
-    page += 1;
-    total = res.data.pages.total;
-    setIsLoaded(false);
-    lock = false;
-    return false;
-  };
-
-  const onIntersect = async ([entry], observer) => {
-    if (entry.isIntersecting && !isLoaded && !lock) {
-      lock = true;
-      setIsLoaded(true);
-      observer.unobserve(entry.target);
-      const result = await getMoreItem();
-      if (!result) observer.observe(entry.target);
-    }
-  };
-
-  useEffect(() => {
-    let observer;
-    if (target) {
-      observer = new IntersectionObserver(onIntersect, {
-        threshold: 0.4,
-      });
-      observer.observe(target);
-    }
-    return () => observer && observer.disconnect();
-  }, [target]);
-
-  const selectMenuHandler = (index) => {
-    setCurrentTab(index);
-  };
 
   return (
     <div className="my-activity_container">
@@ -100,7 +44,7 @@ function MyActivity() {
           <li
             key={el.name}
             className={currentTab === index ? 'submenu focused' : 'submenu'}
-            onClick={() => selectMenuHandler(index)}
+            onClick={() => setCurrentTab(index)}
             aria-hidden="true"
           >
             {el.name}
